+++ env.go
@@ -35,6 +35,10 @@ func NewEnvironment(ctx context.Context) *Environment {
 }
 
 // Context returns the base context of the environment.
+//
+// In almost all cases, you should use Go rather than this to obtain
+// a context because goroutines started by Go can be synchronized by
+// Wait.  Be warned.
 func (e *Environment) Context() context.Context {
 	return e.ctx
 }