+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.1.0] - 2016-08-24
 ### Added
 - `IDGenerator` generates UUID-like ID string for request tracking.
 - `Go` issues new request tracking ID and store it in the derived context.
@@ -24,5 +26,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Use [cybozu-go/netutil](https://github.com/cybozu-go/netutil).
 - Conform to cybozu-go/log v1.1.0 spec.
 
-[Unreleased]: https://github.com/cybozu-go/cmd/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/cybozu-go/cmd/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/cybozu-go/cmd/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/cybozu-go/cmd/compare/v1.0.0...v1.0.1